+++ R/tar_make.R
@@ -14,8 +14,7 @@
 #'   symbols, a character vector, or `tidyselect` helpers like [starts_with()].
 #' @param reporter Character of length 1, name of the reporter to user.
 #'   Controls how messages are printed as targets run in the pipeline.
-#'   Defaults to the `TAR_MAKE_REPORTER` environment variable if set
-#'   and `"verbose"` otherwise. Choices:
+#'   Defaults to `tar_config_get("reporter_make")`. Choices:
 #'   * `"verbose"`: print one message for each target that runs (default).
 #'   * `"silent"`: print nothing.
 #'   * `"timestamp"`: print a time-stamped message for each target that runs.
@@ -42,7 +41,7 @@
 #' })
 tar_make <- function(
   names = NULL,
-  reporter = Sys.getenv("TAR_MAKE_REPORTER", unset = "verbose"),
+  reporter = targets::tar_config_get("reporter_make"),
   callr_function = callr::r,
   callr_arguments = targets::callr_args_default(callr_function, reporter),
   script = targets::tar_config_get("script"),